Tutorial 02 | FastAPI Template | FastAPI Jinja/requirements.txt
aiofiles==0.7.0
alembic==1.6.5
asgiref==3.3.4
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
fastapi==0.65.2
firebase-admin==5.0.1
gcloud==0.17.0
google-api-core==1.30.0
google-api-python-client==2.9.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.0
google-cloud-firestore==2.1.3
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.38.0
h11==0.12.0
httplib2==0.19.1
httptools==0.2.0
idna==2.10
Jinja2==3.0.1
jws==0.1.3
Mako==1.1.4
MarkupSafe==2.0.1
msgpack==1.0.2
oauth2client==3.0.0
packaging==20.9
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pydantic==1.8.2
pyparsing==2.4.7
Pyrebase==3.0.27
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-jwt==2.0.1
python-multipart==0.0.5
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.18
SQLAlchemy-Utils==0.37.6
starlette==0.14.2
text-unidecode==1.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
